internationalization and clearning, also change of name
