--- modules/WebExportersPlugin/src/main/resources/net/clementlevallois/web/exports/plugin/controller/Bundle.properties
@@ -1,23 +1,0 @@
-JPanelWebExport.jTextFieldGithubErrorMsg.text=
-JPanelWebExport.jLabel2.text=7. Everything should be fine now! Move to the "publish" tab of this window and publish your network!
-JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https://github.com/login/device
-JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
-JPanelWebExport.jButtonGithubConnect.text=Connect to Github
-JPanelWebExport.tabGithub.TabConstraints.tabTitle=Set up (to do just once)
-JPanelWebExport.jPanel1.TabConstraints.tabTitle=Publish
-JPanelWebExport.jLabel3.text=1. You need to have a Github Account.
-JPanelWebExport.jLabel4.text=if you don't have one, it is free and takes one minute to create here: 
-JPanelWebExport.jTextField1.text=https://github.com/signup
-JPanelWebExport.jLabel5.text=2. Login to your Github Account
-JPanelWebExport.jLabel6.text=3. Install the Gephi-lite app on Github, just by clicking on "Install" here: 
-JPanelWebExport.jTextField2.text=https://github.com/apps/gephi-lite
-JPanelWebExport.jLabel7.text=4. Start to connect to the Gephi-lite app, click here:
-JPanelWebExport.jLabel8.text=5. Wait until a code appears there and copy it -->
-JPanelWebExport.jLabel9.text=6. Paste the code on this website and click on "continue" then "authorize"
-JPanelWebExport.jLabel1.text=\ if any error occurred, the error messages will appear below
-JPanelWebExport.jLabelAlreadyLoggedIn.text=It seems you already did this set-up, so no need to do it again!
-JPanelWebExport.jButtonResetLogin.text=Reset / erase the login
-JPanelWebExport.jLabel10.text=\ud83d\udd25 Please note that your network will be made public - don't use this feature if your data is private. \ud83d\udd25
-JPanelWebExport.jTextFieldUrlPublishedGraph.text=The link (url) to your network will appear below.
-JPanelWebExport.jButtonPublish.text=Publish your network to the web!
-CTL_PublishWeb=Publish to the web
