+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle.properties
@@ -1,0 +1,42 @@
+#Fri Sep 02 13:17:01 CEST 2022
+JPanelWebExport.jButtonPublish.text=Publish your network to the web\!
+general.message.github.past_code=6. Paste the code on this site and click &quot;continue&quot; then &quot;authorize&quot;\:
+general.message.error.unspecific_error_while_publishing=Error while posting, without specific details.
+general.message.warning_setup_already_done=It looks like you&\'ve already done the setup, so no need to do it again\!
+JPanelWebExport.jLabel3.text=1. You need to have a Github Account.
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+OpenIDE-Module-Long-Description=This plugin converts your network into a gexf file. This file is sent to a "Gist" of Github, where it becomes publicly accessible. Then, a link is created, which allows you to view, explore and share your network with the Retina web application. Be careful, do not publish your networks if the data is confidential\!
+JPanelWebExport.jLabel8.text=5. Wait until a code appears there and copy it -->
+general.message.warning_confidentiality=\ud83d\udd25Warning\! Do not publish your network if it contains confidential data\!\ud83d\udd25
+general.message.error_message=Error message\:
+general.message.success_switch_to_publish=Hit. You can go to "publish".
+general.message.url_published_gexf=Your network URL is\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Login to your Github Account
+general.message.info.url_will_appear_below=\ud83d\udd17 The link (url) of your network will appear below \ud83d\udd17
+general.message.file_sent_from_gephi=File sent from Gephi
+general.message.github.install_gephi_app=if you don&\'t have one, it&\'s free and takes a minute to create here\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Connect to Github
+general.message.error.no_token=Unable to publish. Did you do the settings correctly?
+general.message.success_ready_to_publish=7. Everything should be fine now\! Go to the &quot;publish&quot; tab of this window and publish your network\!
+general.message.error.cant_retrieve_user_code=Unable to retrieve their user code. Are you properly connected to Github?
+general.message.error.no_open_project=You must create or open a project otherwise there is nothing to publish.
+OpenIDE-Module-Name=Publish your network to the web
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Start to connect to the Gephi-lite app, click here\:
+JPanelWebExport.jLabel1.text=\ if any error occurred, the error messages will appear below
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Set up (to do just once)
+general.message.error.gexf_not_retrieved=Unable to create gexf file from your network.
+general.message.success_reset=The reset was successful.
+general.message.error.gist_creation=Error creating online gexf file. Error code\:
+OpenIDE-Module-Short-Description=This plugin allows you to publish, visualize and share your network on the web, directly from Gephi.
+general.message.error.empty_network=Your network is empty. Create a network to publish it.
+general.message.plugin.name=Publish to the web
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Reset / erase the login
+general.message.error.no_user_code=Mistake. The user code has not been validated on the website.
+general.message.url_published_on_retina=The url to view your network directly on Retina is\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Publish
+general.verb.close_window=Close
+general.message.github.create_account=if you don&\'t have one, it&\'s free and takes a minute to create here\:
