+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_ca.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:16:57 CEST 2022
+JPanelWebExport.jButtonPublish.text=Publica la teva xarxa al web\!
+general.message.github.past_code=6. Enganxeu el codi en aquest lloc i feu clic a &quot;continua&quot; i despr\u00e9s a &quot;autoritza&quot;\:
+general.message.error.unspecific_error_while_publishing=Error en publicar, sense detalls espec\u00edfics.
+general.message.warning_setup_already_done=Sembla que ja has fet la configuraci\u00f3, aix\u00ed que no cal que tornis a fer-ho\!
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+JPanelWebExport.jLabel3.text=1. Necessites tenir un compte de Github.
+CTL_PublishWeb=Publicar a la web
+OpenIDE-Module-Long-Description=Aquest connector converteix la vostra xarxa en un fitxer gexf. Aquest fitxer s'envia a un "Gist" de Github, on es fa accessible p\u00fablicament. Aleshores, es crea un enlla\u00e7, que us permet veure, explorar i compartir la vostra xarxa amb l'aplicaci\u00f3 web Retina. Aneu amb compte, no publiqueu les vostres xarxes si les dades s\u00f3n confidencials\!
+JPanelWebExport.jLabel8.text=5. Espereu fins que hi aparegui un codi i copieu-lo -->;
+general.message.warning_confidentiality=\ud83d\udd25Av\u00eds\! No publiquis la teva xarxa si cont\u00e9 dades confidencials\!\ud83d\udd25
+general.message.error_message=Missatge d'error\:
+general.message.success_switch_to_publish=Hit. Podeu anar a "publicar".
+general.message.url_published_gexf=L'URL de la vostra xarxa \u00e9s\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Inicieu sessi\u00f3 al vostre compte de Github
+general.message.info.url_will_appear_below=\ud83d\udd17 L&\'enlla\u00e7 (url) de la teva xarxa apareixer\u00e0 a continuaci\u00f3 \ud83d\udd17
+general.message.file_sent_from_gephi=Fitxer enviat des de Gephi
+general.message.github.install_gephi_app=si no en teniu, \u00e9s gratu\u00eft i triga un minut a crear-lo aqu\u00ed\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Connecteu-vos a Github
+general.message.error.no_token=No es pot publicar. Has fet la configuraci\u00f3 correctament?
+general.message.success_ready_to_publish=7. Ara tot hauria d&\'anar b\u00e9\! Aneu a la pestanya &quot;publicar&quot; d&\'aquesta finestra i publiqueu la vostra xarxa\!
+general.message.error.cant_retrieve_user_code=No es pot recuperar el seu codi d'usuari. Esteu connectat correctament a Github?
+general.message.error.no_open_project=Heu de crear o obrir un projecte en cas contrari no hi ha res a publicar.
+OpenIDE-Module-Name=Publica la teva xarxa al web
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Comenceu a connectar-vos a l'aplicaci\u00f3 Gephi-lite, feu clic aqu\u00ed\:
+JPanelWebExport.jLabel1.text=si s'ha produ\u00eft algun error, els missatges d'error apareixeran a continuaci\u00f3
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Configura (per fer nom\u00e9s una vegada)
+general.message.error.gexf_not_retrieved=No es pot crear un fitxer gexf des de la vostra xarxa.
+general.message.success_reset=El restabliment va tenir \u00e8xit.
+general.message.error.gist_creation=S'ha produ\u00eft un error en crear el fitxer gexf en l\u00ednia. Codi d'error\:
+OpenIDE-Module-Short-Description=Aquest connector us permet publicar, visualitzar i compartir la vostra xarxa al web, directament des de Gephi.
+general.message.error.empty_network=La teva xarxa est\u00e0 buida. Crear una xarxa per publicar-lo.
+general.message.plugin.name=Publicar a la web
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Restableix / esborra l'inici de sessi\u00f3
+general.message.error.no_user_code=Errada. El codi d'usuari no ha estat validat al lloc web.
+general.message.url_published_on_retina=L'URL per veure la vostra xarxa directament a Retina \u00e9s\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Publicar
+general.verb.close_window=Tanca
+general.message.github.create_account=si no en teniu, \u00e9s gratu\u00eft i triga un minut a crear-lo aqu\u00ed\:
