+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_cs.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:16:58 CEST 2022
+JPanelWebExport.jButtonPublish.text=Zve\u0159ejn\u011Bte svou s\u00ED\u0165 na webu\!
+general.message.github.past_code=6. Vlo\u017Ete k\u00F3d na tuto str\u00E1nku a klikn\u011Bte na \u201Epokra\u010Dovat\u201C a pot\u00E9 na \u201Eautorizovat\u201C\:
+general.message.error.unspecific_error_while_publishing=Chyba p\u0159i odes\u00EDl\u00E1n\u00ED bez konkr\u00E9tn\u00EDch podrobnost\u00ED.
+general.message.warning_setup_already_done=Zd\u00E1 se, \u017Ee jste nastaven\u00ED ji\u017E provedli, tak\u017Ee to nen\u00ED t\u0159eba opakovat\!
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+JPanelWebExport.jLabel3.text=1. Mus\u00EDte m\u00EDt \u00FA\u010Det Github.
+CTL_PublishWeb=Publikovat na webu
+OpenIDE-Module-Long-Description=Tento plugin p\u0159evede va\u0161i s\u00ED\u0165 na soubor gexf. Tento soubor je odesl\u00E1n do "Gist" Github, kde se stane ve\u0159ejn\u011B p\u0159\u00EDstupn\u00FDm. Pot\u00E9 se vytvo\u0159\u00ED odkaz, kter\u00FD v\u00E1m umo\u017En\u00ED prohl\u00ED\u017Eet, prozkoum\u00E1vat a sd\u00EDlet va\u0161i s\u00ED\u0165 pomoc\u00ED webov\u00E9 aplikace Retina. Bu\u010Fte opatrn\u00ED, nezve\u0159ej\u0148ujte sv\u00E9 s\u00EDt\u011B, pokud jsou data d\u016Fv\u011Brn\u00E1\!
+JPanelWebExport.jLabel8.text=5. Po\u010Dkejte, a\u017E se tam objev\u00ED k\u00F3d a zkop\u00EDrujte jej -->;
+general.message.warning_confidentiality=\uD83D\uDD25 Upozorn\u011Bn\u00ED\! Nezve\u0159ej\u0148ujte svou s\u00ED\u0165, pokud obsahuje d\u016Fv\u011Brn\u00E1 data\!\uD83D\uDD25
+general.message.error_message=Chybov\u00E9 hl\u00E1\u0161en\u00ED\:
+general.message.success_switch_to_publish=Ude\u0159il. M\u016F\u017Eete p\u0159ej\u00EDt na "publikovat".
+general.message.url_published_gexf=Adresa URL va\u0161\u00ED s\u00EDt\u011B je\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. P\u0159ihlaste se ke sv\u00E9mu \u00FA\u010Dtu Github
+general.message.info.url_will_appear_below=\uD83D\uDD17 N\u00ED\u017Ee se zobraz\u00ED odkaz (url) va\u0161\u00ED s\u00EDt\u011B \uD83D\uDD17
+general.message.file_sent_from_gephi=Soubor odeslan\u00FD z Gephi
+general.message.github.install_gephi_app=pokud \u017E\u00E1dn\u00FD nem\u00E1te, je zdarma a jeho vytvo\u0159en\u00ED zabere minutu zde\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=P\u0159ipojte se ke Github
+general.message.error.no_token=Nelze publikovat. Provedli jste nastaven\u00ED spr\u00E1vn\u011B?
+general.message.success_ready_to_publish=7. U\u017E by m\u011Blo b\u00FDt v\u0161e v po\u0159\u00E1dku\! P\u0159ejd\u011Bte na kartu \u201Epublikovat\u201C v tomto okn\u011B a publikujte svou s\u00ED\u0165\!
+general.message.error.cant_retrieve_user_code=Nelze na\u010D\u00EDst jejich u\u017Eivatelsk\u00FD k\u00F3d. Jste spr\u00E1vn\u011B p\u0159ipojeni ke Github?
+general.message.error.no_open_project=Mus\u00EDte vytvo\u0159it nebo otev\u0159\u00EDt projekt, jinak nen\u00ED co publikovat.
+OpenIDE-Module-Name=Zve\u0159ejn\u011Bte svou s\u00ED\u0165 na webu
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Za\u010Dn\u011Bte se p\u0159ipojovat k aplikaci Gephi-lite, klikn\u011Bte zde\:
+JPanelWebExport.jLabel1.text=pokud dojde k n\u011Bjak\u00E9 chyb\u011B, n\u00ED\u017Ee se zobraz\u00ED chybov\u00E1 hl\u00E1\u0161en\u00ED
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Nastavit (prov\u00E9st pouze jednou)
+general.message.error.gexf_not_retrieved=Nelze vytvo\u0159it soubor gexf z va\u0161\u00ED s\u00EDt\u011B.
+general.message.success_reset=Reset byl \u00FAsp\u011B\u0161n\u00FD.
+general.message.error.gist_creation=Chyba p\u0159i vytv\u00E1\u0159en\u00ED online souboru gexf. Chybov\u00FD k\u00F3d\:
+OpenIDE-Module-Short-Description=Tento plugin v\u00E1m umo\u017E\u0148uje publikovat, vizualizovat a sd\u00EDlet va\u0161i s\u00ED\u0165 na webu p\u0159\u00EDmo z Gephi.
+general.message.error.empty_network=Va\u0161e s\u00ED\u0165 je pr\u00E1zdn\u00E1. Vytvo\u0159te s\u00ED\u0165 pro jej\u00ED publikov\u00E1n\u00ED.
+general.message.plugin.name=Publikovat na webu
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Resetovat / vymazat p\u0159ihl\u00E1\u0161en\u00ED
+general.message.error.no_user_code=Chyba. U\u017Eivatelsk\u00FD k\u00F3d nebyl na webu ov\u011B\u0159en.
+general.message.url_published_on_retina=Adresa URL pro zobrazen\u00ED va\u0161\u00ED s\u00EDt\u011B p\u0159\u00EDmo na Retina je\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Publikovat
+general.verb.close_window=Zav\u0159\u00EDt
+general.message.github.create_account=pokud \u017E\u00E1dn\u00FD nem\u00E1te, je zdarma a jeho vytvo\u0159en\u00ED zabere minutu zde\:
