+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_de.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:00 CEST 2022
+JPanelWebExport.jButtonPublish.text=Ver\u00F6ffentlichen Sie Ihr Netzwerk im Web\!
+general.message.github.past_code=6. F\u00FCgen Sie den Code auf dieser Seite ein und klicken Sie auf \u201EWeiter\u201C und dann auf \u201EAutorisieren\u201C\:
+general.message.error.unspecific_error_while_publishing=Fehler beim Posten, ohne konkrete Details.
+general.message.warning_setup_already_done=Es sieht so aus, als h\u00E4tten Sie die Einrichtung bereits durchgef\u00FChrt, also m\u00FCssen Sie es nicht noch einmal tun\!
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+JPanelWebExport.jLabel3.text=1. Sie ben\u00F6tigen ein Github-Konto.
+CTL_PublishWeb=Im Web ver\u00F6ffentlichen
+OpenIDE-Module-Long-Description=Dieses Plugin konvertiert Ihr Netzwerk in eine gexf-Datei. Diese Datei wird an einen \u201EGist\u201C von Github gesendet, wo sie \u00F6ffentlich zug\u00E4nglich wird. Dann wird ein Link erstellt, mit dem Sie Ihr Netzwerk anzeigen, erkunden und mit der Retina-Webanwendung teilen k\u00F6nnen. Seien Sie vorsichtig, ver\u00F6ffentlichen Sie Ihre Netzwerke nicht, wenn die Daten vertraulich sind\!
+JPanelWebExport.jLabel8.text=5. Warten bis dort ein Code erscheint und kopieren -->;
+general.message.warning_confidentiality=\uD83D\uDD25Achtung\! Ver\u00F6ffentlichen Sie Ihr Netzwerk nicht, wenn es vertrauliche Daten enth\u00E4lt\!\uD83D\uDD25
+general.message.error_message=Fehlermeldung\:
+general.message.success_switch_to_publish=Schlag. Sie k\u00F6nnen auf "Ver\u00F6ffentlichen" gehen.
+general.message.url_published_gexf=Ihre Netzwerk-URL lautet\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Melden Sie sich bei Ihrem Github-Konto an
+general.message.info.url_will_appear_below=\uD83D\uDD17 Der Link (URL) Ihres Netzwerks wird unten angezeigt \uD83D\uDD17
+general.message.file_sent_from_gephi=Von Gephi gesendete Datei
+general.message.github.install_gephi_app=Wenn Sie keins haben, ist es kostenlos und es dauert eine Minute, es hier zu erstellen\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Verbinden Sie sich mit Github
+general.message.error.no_token=Ver\u00F6ffentlichung nicht m\u00F6glich. Hast du die Einstellungen richtig gemacht?
+general.message.success_ready_to_publish=7. Jetzt sollte alles in Ordnung sein\! Gehen Sie zur Registerkarte &quot;Ver\u00F6ffentlichen&quot; dieses Fensters und ver\u00F6ffentlichen Sie Ihr Netzwerk\!
+general.message.error.cant_retrieve_user_code=Der Benutzercode kann nicht abgerufen werden. Sind Sie richtig mit Github verbunden?
+general.message.error.no_open_project=Sie m\u00FCssen ein Projekt erstellen oder \u00F6ffnen, andernfalls gibt es nichts zu ver\u00F6ffentlichen.
+OpenIDE-Module-Name=Ver\u00F6ffentlichen Sie Ihr Netzwerk im Web
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Beginnen Sie, sich mit der Gephi-lite-App zu verbinden, klicken Sie hier\:
+JPanelWebExport.jLabel1.text=Wenn ein Fehler aufgetreten ist, werden die Fehlermeldungen unten angezeigt
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Einrichten (nur einmal zu tun)
+general.message.error.gexf_not_retrieved=Gexf-Datei kann nicht aus Ihrem Netzwerk erstellt werden.
+general.message.success_reset=Das Zur\u00FCcksetzen war erfolgreich.
+general.message.error.gist_creation=Fehler beim Erstellen der Online-Gexf-Datei. Fehlercode\:
+OpenIDE-Module-Short-Description=Mit diesem Plugin k\u00F6nnen Sie Ihr Netzwerk direkt von Gephi aus im Web ver\u00F6ffentlichen, visualisieren und teilen.
+general.message.error.empty_network=Ihr Netzwerk ist leer. Erstellen Sie ein Netzwerk, um es zu ver\u00F6ffentlichen.
+general.message.plugin.name=Im Web ver\u00F6ffentlichen
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Login zur\u00FCcksetzen / l\u00F6schen
+general.message.error.no_user_code=Fehler. Der Benutzercode wurde auf der Website nicht validiert.
+general.message.url_published_on_retina=Die URL, um Ihr Netzwerk direkt auf Retina anzuzeigen, lautet\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Ver\u00F6ffentlichen
+general.verb.close_window=Nah dran
+general.message.github.create_account=Wenn Sie keins haben, ist es kostenlos und es dauert eine Minute, es hier zu erstellen\:
