+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_es.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:03 CEST 2022
+JPanelWebExport.jButtonPublish.text=\u00A1Publique su red en la web\!
+general.message.github.past_code=6. Pegue el c\u00F3digo en este sitio y haga clic en &quot;continuar&quot; y luego en &quot;autorizar&quot;\:
+general.message.error.unspecific_error_while_publishing=Error al publicar, sin detalles espec\u00EDficos.
+general.message.warning_setup_already_done=Parece que ya ha realizado la configuraci\u00F3n, \u00A1as\u00ED que no es necesario que lo haga de nuevo\!
+JPanelWebExport.jTextField1.text=https\://github.com/reg\u00EDstrate
+JPanelWebExport.jLabel3.text=1. Debe tener una cuenta de Github.
+CTL_PublishWeb=Publicar en la web
+OpenIDE-Module-Long-Description=Este complemento convierte su red en un archivo gexf. Este archivo se env\u00EDa a un "Gist" de Github, donde se vuelve accesible p\u00FAblicamente. Luego, se crea un enlace que le permite ver, explorar y compartir su red con la aplicaci\u00F3n web Retina. \u00A1Cuidado, no publiques tus redes si los datos son confidenciales\!
+JPanelWebExport.jLabel8.text=5. Espere hasta que aparezca un c\u00F3digo all\u00ED y c\u00F3pielo -->;
+general.message.warning_confidentiality=\uD83D\uDD25 \u00A1Advertencia\! \u00A1No publiques tu red si contiene datos confidenciales\!\uD83D\uDD25
+general.message.error_message=Mensaje de error\:
+general.message.success_switch_to_publish=Pegar. Puedes ir a "publicar".
+general.message.url_published_gexf=Su URL de red es\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/dispositivo
+JPanelWebExport.jLabel5.text=2. Inicie sesi\u00F3n en su cuenta de Github
+general.message.info.url_will_appear_below=\uD83D\uDD17 A continuaci\u00F3n aparecer\u00E1 el enlace (url) de tu red \uD83D\uDD17
+general.message.file_sent_from_gephi=Archivo enviado desde Gephi
+general.message.github.install_gephi_app=si no tiene uno, es gratis y toma un minuto crearlo aqu\u00ED\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Conectarse a Github
+general.message.error.no_token=No se puede publicar. \u00BFHiciste la configuraci\u00F3n correctamente?
+general.message.success_ready_to_publish=7. \u00A1Todo deber\u00EDa estar bien ahora\! Vaya a la pesta\u00F1a &quot;publicar&quot; de esta ventana y publique su red.
+general.message.error.cant_retrieve_user_code=No se puede recuperar su c\u00F3digo de usuario. \u00BFEst\u00E1s correctamente conectado a Github?
+general.message.error.no_open_project=Debe crear o abrir un proyecto, de lo contrario no hay nada que publicar.
+OpenIDE-Module-Name=Publica tu red en la web
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Comience a conectarse a la aplicaci\u00F3n Gephi-lite, haga clic aqu\u00ED\:
+JPanelWebExport.jLabel1.text=si se produjo alg\u00FAn error, los mensajes de error aparecer\u00E1n a continuaci\u00F3n
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Configurar (para hacer solo una vez)
+general.message.error.gexf_not_retrieved=No se puede crear el archivo gexf desde su red.
+general.message.success_reset=El reinicio fue exitoso.
+general.message.error.gist_creation=Error al crear el archivo gexf en l\u00EDnea. C\u00F3digo de error\:
+OpenIDE-Module-Short-Description=Este complemento le permite publicar, visualizar y compartir su red en la web, directamente desde Gephi.
+general.message.error.empty_network=Su red est\u00E1 vac\u00EDa. Crea una red para publicarlo.
+general.message.plugin.name=Publicar en la web
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Restablecer / borrar el inicio de sesi\u00F3n
+general.message.error.no_user_code=Error. El c\u00F3digo de usuario no ha sido validado en el sitio web.
+general.message.url_published_on_retina=La URL para ver su red directamente en Retina es\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Publicar
+general.verb.close_window=Cerca
+general.message.github.create_account=si no tiene uno, es gratis y toma un minuto crearlo aqu\u00ED\:
