+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_fr.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 10:41:27 CEST 2022
+JPanelWebExport.jButtonPublish.text=Publiez votre r\u00e9seau sur le Web \!
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+JPanelWebExport.jLabel3.text=1. Vous devez avoir un compte Github.
+JPanelWebExport.jLabel8.text=5. Attendez qu'un code y apparaisse et copiez-le -->;
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Connectez-vous \u00e0 votre compte Github
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Connectez-vous \u00e0 Github
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Commencez \u00e0 vous connecter \u00e0 l'application Gephi-lite, cliquez ici \:
+JPanelWebExport.jLabel1.text=si une erreur s'est produite, les messages d'erreur appara\u00eetront ci-dessous
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Mise en place (\u00e0 faire une seule fois)
+general.message.plugin.name=Publier sur le Web
+general.verb.close_window=Fermer
+general.message.success_switch_to_publish=Succ\u00e8s. Vous pouvez aller sur "publier".
+general.message.github.create_account=si vous n'en avez pas, c'est gratuit et cela prend une minute \u00e0 cr\u00e9er ici :
+general.message.github.install_gephi_app=si vous n'en avez pas, c'est gratuit et cela prend une minute \u00e0 cr\u00e9er ici :
+general.message.github.past_code=6. Collez le code sur ce site et cliquez sur "continuer" puis "autoriser" :
+general.message.success_reset=La r\u00e9initialisation a \u00e9t\u00e9 effectu\u00e9e avec succ\u00e8s.
+general.message.success_ready_to_publish=7. Tout devrait bien se passer maintenant ! Allez dans l'onglet "publier" de cette fen\u00eatre et publiez votre r\u00e9seau !
+general.message.error.no_user_code=Erreur. Le code utilisateur n'a pas \u00e9t\u00e9 valid\u00e9 sur le site web.
+general.message.error.cant_retrieve_user_code=Impossible de r\u00e9cup\u00e9rer leur code utilisateur. Etes vous bien connect\u00e9 \u00e0 Github ?
+general.message.error.gexf_not_retrieved=Impossible de cr\u00e9er le fichier gexf \u00e0 partir de votre r\u00e9seau.
+general.message.error.unspecific_error_while_publishing=Erreur lors de la publication, sans d\u00e9tail sp\u00e9cifique.
+general.message.error.no_token=Impossible de proc\u00e9der \u00e0 la publication. Avez-vous bien fait le param\u00e9trage ?
+general.message.error.gist_creation=Erreur lors de la cr\u00e9ation du fichier gexf en ligne. Code d'erreur : 
+general.message.error.no_open_project=Vous devez cr\u00e9er ou ouvrir un projet sinon il n'y a rien \u00e0 publier.
+general.message.error.empty_network=Votre r\u00e9seau est vide. Cr\u00e9ez un r\u00e9seau pour le publier.
+general.message.error_message=Message d'erreur: 
+general.message.warning_confidentiality=\ud83d\udd25Attention! Ne publiez pas votre r\u00e9seau s'il contient des donn\u00e9es confidentielles!
+general.message.warning_setup_already_done=Il semble que vous ayez d\u00e9j\u00e0 fait la configuration, donc pas besoin de la refaire \!
+general.message.info.url_will_appear_below= \ud83d\udd17 Le lien (url) de votre r\u00e9seau va appara\u00eetre ci-dessous
+general.message.url_published_gexf=L'url de votre r\u00e9seau est: 
+general.message.file_sent_from_gephi=Fichier envoy\u00e9 depuis Gephi
+general.message.url_published_on_retina=L'url pour visualiser votre r\u00e9seau directement sur Retina est: 
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=R\u00e9initialiser / effacer la connexion
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Publier
+OpenIDE-Module-Name: Publiez votre r\u00e9seau sur le web
+OpenIDE-Module-Short-Description: Ce plugin vous permet de publier, visualiser et partager votre r\u00e9seau sur le web, directement \u00e0 partir de Gephi.
+OpenIDE-Module-Long-Description: Ce plugin convertit votre r\u00e9seau en un fichier gexf. Ce fichier est envoy\u00e9 sur un "Gist" de Github, o\u00f9 il devient accessible publiquement. Puis, un lien est cr\u00e9\u00e9, qui vous permet de visualiser, explorer et partager votre r\u00e9seau avec l'application web Retina. Attention, ne publiez pas vos r\u00e9seaux si les donn\u00e9es sont confidentielles !
+
