+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_it.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:05 CEST 2022
+JPanelWebExport.jButtonPublish.text=Pubblica la tua rete sul web\!
+general.message.github.past_code=6. Incolla il codice su questo sito e clicca su &quot;continua&quot; poi &quot;autorizza&quot;\:
+general.message.error.unspecific_error_while_publishing=Errore durante la pubblicazione, senza dettagli specifici.
+general.message.warning_setup_already_done=Sembra che tu abbia gi\u00e0 eseguito l&\'installazione, quindi non c&\'\u00e8 bisogno di farlo di nuovo\!
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+JPanelWebExport.jLabel3.text=1. Devi avere un account Github.
+CTL_PublishWeb=Pubblica sul web
+OpenIDE-Module-Long-Description=Questo plugin converte la tua rete in un file gexf. Questo file viene inviato a un "Gist" di Github, dove diventa pubblicamente accessibile. Quindi, viene creato un collegamento che consente di visualizzare, esplorare e condividere la rete con l'applicazione Web Retina. Attenzione, non pubblicare le tue reti se i dati sono riservati\!
+JPanelWebExport.jLabel8.text=5. Attendi che venga visualizzato un codice e copialo -->;
+general.message.warning_confidentiality=\ud83d\udd25Attenzione\! Non pubblicare la tua rete se contiene dati riservati\!\ud83d\udd25
+general.message.error_message=Messaggio di errore\:
+general.message.success_switch_to_publish=Colpo. Puoi andare su "pubblica".
+general.message.url_published_gexf=L'URL della tua rete \u00e8\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Accedi al tuo account Github
+general.message.info.url_will_appear_below=\ud83d\udd17 Il link (url) della tua rete apparir\u00e0 sotto \ud83d\udd17
+general.message.file_sent_from_gephi=File inviato da Gephi
+general.message.github.install_gephi_app=se non ne hai uno, \u00e8 gratuito e richiede un minuto per crearlo qui\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Collegati a Github
+general.message.error.no_token=Impossibile pubblicare. Hai fatto le impostazioni correttamente?
+general.message.success_ready_to_publish=7. Adesso dovrebbe essere tutto a posto\! Vai alla scheda &quot;pubblica&quot; di questa finestra e pubblica la tua rete\!
+general.message.error.cant_retrieve_user_code=Impossibile recuperare il codice utente. Sei connesso correttamente a Github?
+general.message.error.no_open_project=Devi creare o aprire un progetto altrimenti non c'\u00e8 nulla da pubblicare.
+OpenIDE-Module-Name=Pubblica la tua rete sul web
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Inizia a connetterti all'app Gephi-lite, clicca qui\:
+JPanelWebExport.jLabel1.text=se si \u00e8 verificato un errore, i messaggi di errore verranno visualizzati di seguito
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Configurazione (da fare solo una volta)
+general.message.error.gexf_not_retrieved=Impossibile creare il file gexf dalla tua rete.
+general.message.success_reset=Il ripristino \u00e8 riuscito.
+general.message.error.gist_creation=Errore durante la creazione del file gexf online. Codice di errore\:
+OpenIDE-Module-Short-Description=Questo plugin ti permette di pubblicare, visualizzare e condividere la tua rete sul web, direttamente da Gephi.
+general.message.error.empty_network=La tua rete \u00e8 vuota. Crea una rete per pubblicarlo.
+general.message.plugin.name=Pubblica sul web
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Reimposta/cancella il login
+general.message.error.no_user_code=Errore. Il codice utente non \u00e8 stato convalidato sul sito.
+general.message.url_published_on_retina=L'url per visualizzare la tua rete direttamente su Retina \u00e8\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Pubblicare
+general.verb.close_window=Chiudere
+general.message.github.create_account=se non ne hai uno, \u00e8 gratuito e richiede un minuto per crearlo qui\:
