+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_ja.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:08 CEST 2022
+JPanelWebExport.jButtonPublish.text=\u3042\u306A\u305F\u306E\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u3092\u30A6\u30A7\u30D6\u306B\u516C\u958B\u3057\u307E\u3057\u3087\u3046\uFF01
+general.message.github.past_code=6. \u3053\u306E\u30B5\u30A4\u30C8\u306B\u30B3\u30FC\u30C9\u3092\u8CBC\u308A\u4ED8\u3051\u3066\u3001[\u7D9A\u884C]\u3001[\u627F\u8A8D] \u306E\u9806\u306B\u30AF\u30EA\u30C3\u30AF\u3057\u307E\u3059\u3002
+general.message.error.unspecific_error_while_publishing=\u5177\u4F53\u7684\u306A\u8A73\u7D30\u306A\u3057\u3067\u3001\u6295\u7A3F\u4E2D\u306B\u30A8\u30E9\u30FC\u304C\u767A\u751F\u3057\u307E\u3057\u305F\u3002
+general.message.warning_setup_already_done=\u30BB\u30C3\u30C8\u30A2\u30C3\u30D7\u306F\u3059\u3067\u306B\u5B8C\u4E86\u3057\u3066\u3044\u308B\u3088\u3046\u3067\u3059\u306E\u3067\u3001\u518D\u5EA6\u30BB\u30C3\u30C8\u30A2\u30C3\u30D7\u3059\u308B\u5FC5\u8981\u306F\u3042\u308A\u307E\u305B\u3093\u3002
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+JPanelWebExport.jLabel3.text=1. Github \u30A2\u30AB\u30A6\u30F3\u30C8\u304C\u5FC5\u8981\u3067\u3059\u3002
+CTL_PublishWeb=Web \u306B\u516C\u958B\u3059\u308B
+OpenIDE-Module-Long-Description=\u3053\u306E\u30D7\u30E9\u30B0\u30A4\u30F3\u306F\u3001\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u3092 gexf \u30D5\u30A1\u30A4\u30EB\u306B\u5909\u63DB\u3057\u307E\u3059\u3002\u3053\u306E\u30D5\u30A1\u30A4\u30EB\u306F\u3001Github \u306E "Gist" \u306B\u9001\u4FE1\u3055\u308C\u3001\u4E00\u822C\u306B\u30A2\u30AF\u30BB\u30B9\u53EF\u80FD\u306B\u306A\u308A\u307E\u3059\u3002\u6B21\u306B\u3001\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u3092\u8868\u793A\u3001\u63A2\u7D22\u3001Retina Web \u30A2\u30D7\u30EA\u30B1\u30FC\u30B7\u30E7\u30F3\u3068\u5171\u6709\u3067\u304D\u308B\u30EA\u30F3\u30AF\u304C\u4F5C\u6210\u3055\u308C\u307E\u3059\u3002\u30C7\u30FC\u30BF\u304C\u6A5F\u5BC6\u60C5\u5831\u3067\u3042\u308B\u5834\u5408\u306F\u3001\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u3092\u516C\u958B\u3057\u306A\u3044\u3067\u304F\u3060\u3055\u3044\u3002
+JPanelWebExport.jLabel8.text=5.\u305D\u3053\u306B\u30B3\u30FC\u30C9\u304C\u8868\u793A\u3055\u308C\u308B\u307E\u3067\u5F85\u3063\u3066\u304B\u3089\u30B3\u30D4\u30FC\u3057\u307E\u3059 -->;
+general.message.warning_confidentiality=\uD83D\uDD25\u8B66\u544A\uFF01\u6A5F\u5BC6\u30C7\u30FC\u30BF\u304C\u542B\u307E\u308C\u3066\u3044\u308B\u5834\u5408\u306F\u3001\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u3092\u516C\u958B\u3057\u306A\u3044\u3067\u304F\u3060\u3055\u3044\!\uD83D\uDD25
+general.message.error_message=\u30A8\u30E9\u30FC\u30E1\u30C3\u30BB\u30FC\u30B8\uFF1A
+general.message.success_switch_to_publish=\u6253\u3064\u3002 \u300C\u516C\u958B\u300D\u306B\u79FB\u52D5\u3067\u304D\u307E\u3059\u3002
+general.message.url_published_gexf=\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF URL \u306F\u6B21\u306E\u3068\u304A\u308A\u3067\u3059\u3002
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2.Github \u30A2\u30AB\u30A6\u30F3\u30C8\u306B\u30ED\u30B0\u30A4\u30F3\u3057\u307E\u3059\u3002
+general.message.info.url_will_appear_below=\uD83D\uDD17 \u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u306E\u30EA\u30F3\u30AF (URL) \u304C\u4E0B\u306B\u8868\u793A\u3055\u308C\u307E\u3059 \uD83D\uDD17
+general.message.file_sent_from_gephi=Gephi \u304B\u3089\u9001\u4FE1\u3055\u308C\u305F\u30D5\u30A1\u30A4\u30EB
+general.message.github.install_gephi_app=\u304A\u6301\u3061\u3067\u306A\u3044\u5834\u5408\u306F\u7121\u6599\u3067\u3001\u3053\u3053\u3067\u4F5C\u6210\u3059\u308B\u306E\u306B 1 \u5206\u307B\u3069\u304B\u304B\u308A\u307E\u3059\u3002
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Github \u306B\u63A5\u7D9A\u3059\u308B
+general.message.error.no_token=\u516C\u958B\u3067\u304D\u307E\u305B\u3093\u3002\u8A2D\u5B9A\u306F\u6B63\u3057\u304F\u884C\u3044\u307E\u3057\u305F\u304B\uFF1F
+general.message.success_ready_to_publish=7. \u3053\u308C\u3067\u3059\u3079\u3066\u304C\u3046\u307E\u304F\u3044\u304F\u306F\u305A\u3067\u3059\!\u3053\u306E\u30A6\u30A3\u30F3\u30C9\u30A6\u306E\u300C\u516C\u958B\u300D\u30BF\u30D6\u306B\u79FB\u52D5\u3057\u3066\u3001\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u3092\u516C\u958B\u3057\u3066\u304F\u3060\u3055\u3044\u3002
+general.message.error.cant_retrieve_user_code=\u30E6\u30FC\u30B6\u30FC \u30B3\u30FC\u30C9\u3092\u53D6\u5F97\u3067\u304D\u307E\u305B\u3093\u3002 Github \u306B\u6B63\u3057\u304F\u63A5\u7D9A\u3055\u308C\u3066\u3044\u307E\u3059\u304B?
+general.message.error.no_open_project=\u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u3092\u4F5C\u6210\u3059\u308B\u304B\u958B\u304F\u5FC5\u8981\u304C\u3042\u308A\u307E\u3059\u3002\u305D\u3046\u3057\u306A\u3044\u3068\u3001\u516C\u958B\u3059\u308B\u3082\u306E\u304C\u3042\u308A\u307E\u305B\u3093\u3002
+OpenIDE-Module-Name=\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u3092 Web \u306B\u516C\u958B\u3059\u308B
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Gephi-lite \u30A2\u30D7\u30EA\u3078\u306E\u63A5\u7D9A\u3092\u958B\u59CB\u3059\u308B\u306B\u306F\u3001\u3053\u3053\u3092\u30AF\u30EA\u30C3\u30AF\u3057\u307E\u3059\u3002
+JPanelWebExport.jLabel1.text=\u30A8\u30E9\u30FC\u304C\u767A\u751F\u3057\u305F\u5834\u5408\u3001\u30A8\u30E9\u30FC\u30E1\u30C3\u30BB\u30FC\u30B8\u304C\u4E0B\u306B\u8868\u793A\u3055\u308C\u307E\u3059
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=\u30BB\u30C3\u30C8\u30A2\u30C3\u30D7\uFF08\u4E00\u5EA6\u3060\u3051\u884C\u3046\uFF09
+general.message.error.gexf_not_retrieved=\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u304B\u3089 gexf \u30D5\u30A1\u30A4\u30EB\u3092\u4F5C\u6210\u3067\u304D\u307E\u305B\u3093\u3002
+general.message.success_reset=\u30EA\u30BB\u30C3\u30C8\u306F\u6210\u529F\u3057\u307E\u3057\u305F\u3002
+general.message.error.gist_creation=\u30AA\u30F3\u30E9\u30A4\u30F3 gexf \u30D5\u30A1\u30A4\u30EB\u306E\u4F5C\u6210\u4E2D\u306B\u30A8\u30E9\u30FC\u304C\u767A\u751F\u3057\u307E\u3057\u305F\u3002\u30A8\u30E9\u30FC\u30B3\u30FC\u30C9\uFF1A
+OpenIDE-Module-Short-Description=\u3053\u306E\u30D7\u30E9\u30B0\u30A4\u30F3\u3092\u4F7F\u7528\u3059\u308B\u3068\u3001\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u3092 Gephi \u304B\u3089\u76F4\u63A5 Web \u4E0A\u3067\u516C\u958B\u3001\u8996\u899A\u5316\u3001\u5171\u6709\u3067\u304D\u307E\u3059\u3002
+general.message.error.empty_network=\u3042\u306A\u305F\u306E\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u306F\u7A7A\u3067\u3059\u3002\u516C\u958B\u3059\u308B\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u3092\u4F5C\u6210\u3057\u307E\u3059\u3002
+general.message.plugin.name=Web \u306B\u516C\u958B\u3059\u308B
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=\u30ED\u30B0\u30A4\u30F3\u306E\u30EA\u30BB\u30C3\u30C8/\u6D88\u53BB
+general.message.error.no_user_code=\u9593\u9055\u3044\u3002\u30E6\u30FC\u30B6\u30FC \u30B3\u30FC\u30C9\u306F Web \u30B5\u30A4\u30C8\u3067\u691C\u8A3C\u3055\u308C\u3066\u3044\u307E\u305B\u3093\u3002
+general.message.url_published_on_retina=\u30CD\u30C3\u30C8\u30EF\u30FC\u30AF\u3092 Retina \u3067\u76F4\u63A5\u8868\u793A\u3059\u308B\u305F\u3081\u306E URL \u306F\u6B21\u306E\u3068\u304A\u308A\u3067\u3059\u3002
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=\u516C\u958B
+general.verb.close_window=\u8FD1\u3044
+general.message.github.create_account=\u304A\u6301\u3061\u3067\u306A\u3044\u5834\u5408\u306F\u7121\u6599\u3067\u3001\u3053\u3053\u3067\u4F5C\u6210\u3059\u308B\u306E\u306B 1 \u5206\u307B\u3069\u304B\u304B\u308A\u307E\u3059\u3002
