+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_nl.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:09 CEST 2022
+JPanelWebExport.jButtonPublish.text=Publiceer uw netwerk op het web\!
+general.message.github.past_code=6. Plak de code op deze site en klik op &quot;doorgaan&quot; en vervolgens op &quot;autoriseren&quot;\:
+general.message.error.unspecific_error_while_publishing=Fout tijdens het posten, zonder specifieke details.
+general.message.warning_setup_already_done=Het lijkt erop dat je de installatie al hebt gedaan, dus je hoeft het niet nog een keer te doen\!
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+JPanelWebExport.jLabel3.text=1. U moet een Github-account hebben.
+CTL_PublishWeb=Publiceren op internet
+OpenIDE-Module-Long-Description=Deze plug-in converteert uw netwerk naar een gexf-bestand. Dit bestand wordt naar een "Gist" van Github gestuurd, waar het openbaar toegankelijk wordt. Vervolgens wordt een koppeling gemaakt waarmee u uw netwerk kunt bekijken, verkennen en delen met de Retina-webapplicatie. Wees voorzichtig, publiceer uw netwerken niet als de gegevens vertrouwelijk zijn\!
+JPanelWebExport.jLabel8.text=5. Wacht tot daar een code verschijnt en kopieer deze -->;
+general.message.warning_confidentiality=\uD83D\uDD25Waarschuwing\! Publiceer uw netwerk niet als het vertrouwelijke gegevens bevat\!\uD83D\uDD25
+general.message.error_message=Foutmelding\:
+general.message.success_switch_to_publish=Raken. U kunt naar "publiceren" gaan.
+general.message.url_published_gexf=Uw netwerk-URL is\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Log in op uw Github-account
+general.message.info.url_will_appear_below=\uD83D\uDD17 De link (url) van je netwerk verschijnt hieronder \uD83D\uDD17
+general.message.file_sent_from_gephi=Bestand verzonden vanuit Gephi
+general.message.github.install_gephi_app=als je er geen hebt, is het gratis en duurt het even om hier te maken\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Verbinding maken met Github
+general.message.error.no_token=Kan niet publiceren. Heb je de instellingen goed gedaan?
+general.message.success_ready_to_publish=7. Alles zou nu in orde moeten zijn\! Ga naar het tabblad &quot;publiceren&quot; van dit venster en publiceer uw netwerk\!
+general.message.error.cant_retrieve_user_code=Kan hun gebruikerscode niet ophalen. Ben je goed verbonden met Github?
+general.message.error.no_open_project=Je moet een project aanmaken of openen, anders valt er niets te publiceren.
+OpenIDE-Module-Name=Publiceer uw netwerk op internet
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Start om verbinding te maken met de Gephi-lite app, klik hier\:
+JPanelWebExport.jLabel1.text=als er een fout is opgetreden, verschijnen de foutmeldingen hieronder
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Instellen (eenmalig doen)
+general.message.error.gexf_not_retrieved=Kan geen gexf-bestand van uw netwerk maken.
+general.message.success_reset=Het resetten is gelukt.
+general.message.error.gist_creation=Fout bij het maken van online gexf-bestand. Foutcode\:
+OpenIDE-Module-Short-Description=Met deze plug-in kunt u uw netwerk rechtstreeks vanuit Gephi op het web publiceren, visualiseren en delen.
+general.message.error.empty_network=Je netwerk is leeg. Maak een netwerk om het te publiceren.
+general.message.plugin.name=Publiceren op internet
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Reset / wis de login
+general.message.error.no_user_code=Vergissing. De gebruikerscode is niet gevalideerd op de website.
+general.message.url_published_on_retina=De url om uw netwerk direct op Retina te bekijken is\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Publiceren
+general.verb.close_window=sluiten
+general.message.github.create_account=als je er geen hebt, is het gratis en duurt het even om hier te maken\:
