+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_no.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:10 CEST 2022
+JPanelWebExport.jButtonPublish.text=Publiser nettverket ditt p\u00E5 nettet\!
+general.message.github.past_code=6. Lim inn koden p\u00E5 dette nettstedet og klikk &quot;fortsett&quot; og deretter &quot;autoriser&quot;\:
+general.message.error.unspecific_error_while_publishing=Feil under innlegging, uten spesifikke detaljer.
+general.message.warning_setup_already_done=Det ser ut til at du allerede har gjort oppsettet, s\u00E5 du trenger ikke \u00E5 gj\u00F8re det igjen\!
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+JPanelWebExport.jLabel3.text=1. Du m\u00E5 ha en Github-konto.
+CTL_PublishWeb=Publiser p\u00E5 nettet
+OpenIDE-Module-Long-Description=Denne plugin-en konverterer nettverket ditt til en gexf-fil. Denne filen sendes til en "Gist" av Github, hvor den blir offentlig tilgjengelig. Deretter opprettes en lenke som lar deg se, utforske og dele nettverket ditt med Retina-nettapplikasjonen. V\u00E6r forsiktig, ikke publiser nettverkene dine hvis dataene er konfidensielle\!
+JPanelWebExport.jLabel8.text=5. Vent til en kode vises der og kopier den -->;
+general.message.warning_confidentiality=\uD83D\uDD25Advarsel\! Ikke publiser nettverket ditt hvis det inneholder konfidensielle data\!\uD83D\uDD25
+general.message.error_message=Feilmelding\:
+general.message.success_switch_to_publish=Truffet. Du kan g\u00E5 til "publiser".
+general.message.url_published_gexf=Nettverksadressen din er\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Logg p\u00E5 Github-kontoen din
+general.message.info.url_will_appear_below=\uD83D\uDD17 Linken (url) til nettverket ditt vises nedenfor \uD83D\uDD17
+general.message.file_sent_from_gephi=Fil sendt fra Gephi
+general.message.github.install_gephi_app=hvis du ikke har en, er den gratis og tar et minutt \u00E5 lage her\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Koble til Github
+general.message.error.no_token=Kan ikke publisere. Gjorde du innstillingene riktig?
+general.message.success_ready_to_publish=7. Alt skal v\u00E6re bra n\u00E5\! G\u00E5 til &quot;publiser&quot;-fanen i dette vinduet og publiser nettverket ditt\!
+general.message.error.cant_retrieve_user_code=Kan ikke hente brukerkoden deres. Er du riktig koblet til Github?
+general.message.error.no_open_project=Du m\u00E5 opprette eller \u00E5pne et prosjekt ellers er det ingenting \u00E5 publisere.
+OpenIDE-Module-Name=Publiser nettverket ditt p\u00E5 nettet
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Begynn \u00E5 koble til Gephi-lite-appen, klikk her\:
+JPanelWebExport.jLabel1.text=hvis det oppstod en feil, vil feilmeldingene vises nedenfor
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Konfigurer (\u00E5 gj\u00F8re bare \u00E9n gang)
+general.message.error.gexf_not_retrieved=Kan ikke opprette gexf-fil fra nettverket ditt.
+general.message.success_reset=Tilbakestillingen var vellykket.
+general.message.error.gist_creation=Feil ved opprettelse av online gexf-fil. Feil kode\:
+OpenIDE-Module-Short-Description=Denne pluginen lar deg publisere, visualisere og dele nettverket ditt p\u00E5 nettet, direkte fra Gephi.
+general.message.error.empty_network=Nettverket ditt er tomt. Opprett et nettverk for \u00E5 publisere det.
+general.message.plugin.name=Publiser p\u00E5 nettet
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Tilbakestill / slett p\u00E5loggingen
+general.message.error.no_user_code=Feil. Brukerkoden er ikke validert p\u00E5 nettsiden.
+general.message.url_published_on_retina=Nettadressen for \u00E5 se nettverket ditt direkte p\u00E5 Retina er\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=publisere
+general.verb.close_window=Lukk
+general.message.github.create_account=hvis du ikke har en, er den gratis og tar et minutt \u00E5 lage her\:
