+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_pl.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:12 CEST 2022
+JPanelWebExport.jButtonPublish.text=Opublikuj swoj\u0105 sie\u0107 w sieci\!
+general.message.github.past_code=6. Wklej kod na tej stronie i kliknij \u201Ekontynuuj\u201D, a nast\u0119pnie \u201Eautoryzuj\u201D\:
+general.message.error.unspecific_error_while_publishing=B\u0142\u0105d podczas publikowania, bez szczeg\u00F3\u0142owych informacji.
+general.message.warning_setup_already_done=Wygl\u0105da na to, \u017Ce ju\u017C wykona\u0142e\u015B konfiguracj\u0119, wi\u0119c nie musisz tego robi\u0107 ponownie\!
+JPanelWebExport.jLabel3.text=1. Musisz mie\u0107 konto Github.
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+CTL_PublishWeb=Opublikuj w sieci
+OpenIDE-Module-Long-Description=Ta wtyczka konwertuje twoj\u0105 sie\u0107 do pliku gexf. Ten plik jest wysy\u0142any do \u201EGist\u201D Github, gdzie staje si\u0119 publicznie dost\u0119pny. Nast\u0119pnie tworzone jest \u0142\u0105cze, kt\u00F3re umo\u017Cliwia przegl\u0105danie, eksplorowanie i udost\u0119pnianie sieci za pomoc\u0105 aplikacji internetowej Retina. Uwa\u017Caj, nie publikuj swoich sieci, je\u015Bli dane s\u0105 poufne\!
+JPanelWebExport.jLabel8.text=5. Poczekaj, a\u017C pojawi si\u0119 tam kod i skopiuj go -->;
+general.message.warning_confidentiality=\uD83D\uDD25Ostrze\u017Cenie\! Nie publikuj swojej sieci, je\u015Bli zawiera poufne dane\!\uD83D\uDD25
+general.message.error_message=Komunikat o b\u0142\u0119dzie\:
+general.message.success_switch_to_publish=Uderzy\u0107. Mo\u017Cesz przej\u015B\u0107 do "publikowania".
+general.message.url_published_gexf=Adres URL Twojej sieci to\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Zaloguj si\u0119 do swojego konta Github
+general.message.info.url_will_appear_below=\uD83D\uDD17 Link (URL) Twojej sieci pojawi si\u0119 poni\u017Cej \uD83D\uDD17
+general.message.file_sent_from_gephi=Plik wys\u0142any z Gephi
+general.message.github.install_gephi_app=je\u015Bli go nie masz, to nic nie kosztuje i zajmie Ci minut\u0119, aby utworzy\u0107 tutaj\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Po\u0142\u0105cz si\u0119 z Github
+general.message.error.no_token=Nie mo\u017Cna opublikowa\u0107. Czy poprawnie wykona\u0142e\u015B ustawienia?
+general.message.success_ready_to_publish=7. Teraz wszystko powinno by\u0107 w porz\u0105dku\! Przejd\u017A do zak\u0142adki \u201Eopublikuj\u201D w tym oknie i opublikuj swoj\u0105 sie\u0107\!
+general.message.error.cant_retrieve_user_code=Nie mo\u017Cna pobra\u0107 kodu u\u017Cytkownika. Czy masz prawid\u0142owe po\u0142\u0105czenie z Github?
+general.message.error.no_open_project=Musisz utworzy\u0107 lub otworzy\u0107 projekt, w przeciwnym razie nie ma nic do opublikowania.
+OpenIDE-Module-Name=Opublikuj swoj\u0105 sie\u0107 w sieci
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Zacznij \u0142\u0105czy\u0107 si\u0119 z aplikacj\u0105 Gephi-lite, kliknij tutaj\:
+JPanelWebExport.jLabel1.text=je\u015Bli wyst\u0105pi jakikolwiek b\u0142\u0105d, komunikaty o b\u0142\u0119dach pojawi\u0105 si\u0119 poni\u017Cej
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Skonfiguruj (do zrobienia tylko raz)
+general.message.error.gexf_not_retrieved=Nie mo\u017Cna utworzy\u0107 pliku gexf z Twojej sieci.
+general.message.success_reset=Reset si\u0119 powi\u00F3d\u0142.
+general.message.error.gist_creation=B\u0142\u0105d podczas tworzenia pliku gexf online. Kod b\u0142\u0119du\:
+OpenIDE-Module-Short-Description=Ta wtyczka umo\u017Cliwia publikowanie, wizualizacj\u0119 i udost\u0119pnianie sieci w Internecie bezpo\u015Brednio z Gephi.
+general.message.error.empty_network=Twoja sie\u0107 jest pusta. Utw\u00F3rz sie\u0107, aby j\u0105 opublikowa\u0107.
+general.message.plugin.name=Opublikuj w sieci
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Zresetuj / usu\u0144 login
+general.message.error.no_user_code=B\u0142\u0105d. Kod u\u017Cytkownika nie zosta\u0142 zweryfikowany na stronie.
+general.message.url_published_on_retina=Adres URL do przegl\u0105dania sieci bezpo\u015Brednio w Retina to\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Publikowa\u0107
+general.verb.close_window=Blisko
+general.message.github.create_account=je\u015Bli go nie masz, to nic nie kosztuje i zajmie Ci minut\u0119, aby utworzy\u0107 tutaj\:
