+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_pt_BR.properties
@@ -1,0 +1,42 @@
+#Fri Sep 02 13:17:56 CEST 2022
+JPanelWebExport.jButtonPublish.text=Publique sua rede na web\!
+general.message.github.past_code=6. Cole o c\u00F3digo neste site e clique em "continuar" e depois em "autorizar" \:
+general.message.error.unspecific_error_while_publishing=Erro na publica\u00E7\u00E3o, sem detalhes espec\u00EDficos.
+general.message.warning_setup_already_done=Parece que voc\u00EA j\u00E1 fez a configura\u00E7\u00E3o, ent\u00E3o n\u00E3o h\u00E1 necessidade de fazer isso novamente\!
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+JPanelWebExport.jLabel3.text=1. Voc\u00EA deve ter uma conta no Github.
+OpenIDE-Module-Long-Description=Este plugin converte sua rede em um arquivo gexf. Este arquivo \u00E9 carregado em um "Gist" no Github, onde se torna acess\u00EDvel ao p\u00FAblico. Em seguida, \u00E9 criado um link que permite visualizar, explorar e compartilhar sua rede com a aplica\u00E7\u00E3o web Retina. Tenha cuidado, n\u00E3o publique suas redes se os dados forem confidenciais\!
+JPanelWebExport.jLabel8.text=5. Aguarde que um c\u00F3digo apare\u00E7a e copie-o -->;
+general.message.warning_confidentiality=\uD83D\uDD25Caution\! N\u00E3o publique sua rede se ela contiver dados confidenciais\!
+general.message.error_message=Mensagem de erro\: 
+general.message.success_switch_to_publish=Sucesso. Voc\u00EA pode ir para "publicar".
+general.message.url_published_gexf=A url de sua rede \u00E9\: 
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Fa\u00E7a login em sua conta Github
+general.message.info.url_will_appear_below=\uD83D\uDD17 O link (url) para sua rede aparecer\u00E1 abaixo \uD83D\uDD17
+general.message.file_sent_from_gephi=Arquivo enviado por Gephi
+general.message.github.install_gephi_app=se voc\u00EA n\u00E3o tiver, \u00E9 gr\u00E1tis e leva um minuto para criar aqui\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Login para Github
+general.message.error.no_token=Incapaz de publicar. Voc\u00EA j\u00E1 fez as configura\u00E7\u00F5es corretamente?
+general.message.success_ready_to_publish=7. Tudo deve ficar bem agora\! V\u00E1 para a aba "publicar" nesta janela e publique sua rede\!
+general.message.error.cant_retrieve_user_code=N\u00E3o consigo recuperar o c\u00F3digo de usu\u00E1rio deles. Voc\u00EA est\u00E1 conectado ao Github?
+general.message.error.no_open_project=Voc\u00EA deve criar ou abrir um projeto, caso contr\u00E1rio n\u00E3o h\u00E1 nada para publicar.
+OpenIDE-Module-Name=Publique sua rede na web
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Comece a conex\u00E3o com a aplica\u00E7\u00E3o Gephi-lite, clique aqui\:
+JPanelWebExport.jLabel1.text=caso tenha ocorrido um erro, aparecer\u00E3o as seguintes mensagens de erro
+general.message.error.gexf_not_retrieved=Incapaz de criar o arquivo gexf a partir de sua rede.
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Montado (apenas uma vez)
+general.message.success_reset=O reset foi um sucesso.
+general.message.error.gist_creation=Erro ao criar o arquivo gexf on-line. C\u00F3digo de erro \: 
+OpenIDE-Module-Short-Description=Este plugin lhe permite publicar, visualizar e compartilhar sua rede na web, diretamente da Gephi.
+general.message.error.empty_network=Sua rede est\u00E1 vazia. Criar uma rede para public\u00E1-la.
+general.message.plugin.name=Publica\u00E7\u00E3o na Web
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+general.message.error.no_user_code=Erro. O c\u00F3digo do usu\u00E1rio n\u00E3o foi validado no site.
+general.message.url_published_on_retina=A url para visualizar sua rede diretamente na Retina \u00E9\: 
+JPanelWebExport.jButtonResetLogin.text=Restabelecer / apagar conex\u00E3o
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Publicar
+general.verb.close_window=Fechar
+general.message.github.create_account=se voc\u00EA n\u00E3o tiver, \u00E9 gr\u00E1tis e leva um minuto para criar aqui\:
