+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_ro.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:13 CEST 2022
+JPanelWebExport.jButtonPublish.text=Publica\u021Bi-v\u0103 re\u021Beaua pe web\!
+general.message.github.past_code=6. Lipi\u021Bi codul pe acest site \u0219i face\u021Bi clic pe \u201Econtinua\u021Bi\u201D apoi \u201Eautoriza\u021Bi\u201D\:
+general.message.error.unspecific_error_while_publishing=Eroare la postare, f\u0103r\u0103 detalii specifice.
+general.message.warning_setup_already_done=Se pare c\u0103 a\u021Bi f\u0103cut deja configurarea, a\u0219a c\u0103 nu este nevoie s\u0103 o face\u021Bi din nou\!
+JPanelWebExport.jLabel3.text=1. Trebuie s\u0103 ave\u021Bi un cont Github.
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+CTL_PublishWeb=Publica\u021Bi pe web
+OpenIDE-Module-Long-Description=Acest plugin v\u0103 transform\u0103 re\u021Beaua \u00EEntr-un fi\u0219ier gexf. Acest fi\u0219ier este trimis la un \u201EGist\u201D al Github, unde devine accesibil public. Apoi, se creeaz\u0103 un link, care v\u0103 permite s\u0103 vizualiza\u021Bi, s\u0103 explora\u021Bi \u0219i s\u0103 partaja\u021Bi re\u021Beaua dvs. cu aplica\u021Bia web Retina. Aten\u021Bie, nu v\u0103 publica\u021Bi re\u021Belele dac\u0103 datele sunt confiden\u021Biale\!
+JPanelWebExport.jLabel8.text=5. A\u0219tepta\u021Bi p\u00E2n\u0103 c\u00E2nd apare un cod acolo \u0219i copia\u021Bi-l -->;
+general.message.warning_confidentiality=\uD83D\uDD25Aten\u021Bie\! Nu v\u0103 publica\u021Bi re\u021Beaua dac\u0103 con\u021Bine date confiden\u021Biale\!\uD83D\uDD25
+general.message.error_message=Mesaj de eroare\:
+general.message.success_switch_to_publish=Lovit. Pute\u021Bi merge la \u201Epublicare\u201D.
+general.message.url_published_gexf=Adresa URL a re\u021Belei dvs. este\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Conecta\u021Bi-v\u0103 la contul dvs. Github
+general.message.info.url_will_appear_below=\uD83D\uDD17 Link-ul (url-ul) re\u021Belei tale va ap\u0103rea mai jos \uD83D\uDD17
+general.message.file_sent_from_gephi=Fi\u0219ier trimis de la Gephi
+general.message.github.install_gephi_app=dac\u0103 nu ave\u021Bi unul, este gratuit \u0219i dureaz\u0103 un minut pentru a crea aici\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Conecta\u021Bi-v\u0103 la Github
+general.message.error.no_token=Nu se poate publica. Ai facut corect setarile?
+general.message.success_ready_to_publish=7. Totul ar trebui s\u0103 fie bine acum\! Accesa\u021Bi fila \u201Epublicare\u201D din aceast\u0103 fereastr\u0103 \u0219i publica\u021Bi-v\u0103 re\u021Beaua\!
+general.message.error.cant_retrieve_user_code=Nu se poate recupera codul de utilizator. E\u0219ti conectat corect la Github?
+general.message.error.no_open_project=Trebuie s\u0103 crea\u021Bi sau s\u0103 deschide\u021Bi un proiect, altfel nu exist\u0103 nimic de publicat.
+OpenIDE-Module-Name=Publica\u021Bi-v\u0103 re\u021Beaua pe web
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. \u00CEncepe\u021Bi s\u0103 v\u0103 conecta\u021Bi la aplica\u021Bia Gephi-lite, face\u021Bi clic aici\:
+JPanelWebExport.jLabel1.text=dac\u0103 a ap\u0103rut vreo eroare, mesajele de eroare vor ap\u0103rea mai jos
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Configura\u021Bi (a face doar o dat\u0103)
+general.message.error.gexf_not_retrieved=Nu se poate crea fi\u0219ierul gexf din re\u021Beaua dvs.
+general.message.success_reset=Resetarea a avut succes.
+general.message.error.gist_creation=Eroare la crearea fi\u0219ierului gexf online. Cod de eroare\:
+OpenIDE-Module-Short-Description=Acest plugin v\u0103 permite s\u0103 publica\u021Bi, s\u0103 vizualiza\u021Bi \u0219i s\u0103 partaja\u021Bi re\u021Beaua dvs. pe web, direct din Gephi.
+general.message.error.empty_network=Re\u021Beaua dvs. este goal\u0103. Crea\u021Bi o re\u021Bea pentru a o publica.
+general.message.plugin.name=Publica\u021Bi pe web
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Reseta\u021Bi / \u0219terge\u021Bi autentificarea
+general.message.error.no_user_code=Gre\u015Feal\u0103. Codul utilizatorului nu a fost validat pe site.
+general.message.url_published_on_retina=Adresa URL pentru a v\u0103 vizualiza re\u021Beaua direct pe Retina este\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Publica
+general.verb.close_window=\u00CEnchide
+general.message.github.create_account=dac\u0103 nu ave\u021Bi unul, este gratuit \u0219i dureaz\u0103 un minut pentru a crea aici\:
