+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_tr.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:16 CEST 2022
+JPanelWebExport.jButtonPublish.text=A\u011F\u0131n\u0131z\u0131 web'de yay\u0131nlay\u0131n\!
+general.message.github.past_code=6. Kodu bu siteye yap\u0131\u015Ft\u0131r\u0131n ve &quot;devam et&quot;e ve ard\u0131ndan &quot;yetkilendir&quot;e t\u0131klay\u0131n\:
+general.message.error.unspecific_error_while_publishing=Belirli ayr\u0131nt\u0131lar olmadan g\u00F6nderirken hata olu\u015Ftu.
+general.message.warning_setup_already_done=Kurulumu zaten yapm\u0131\u015Fs\u0131n\u0131z gibi g\u00F6r\u00FCn\u00FCyor, bu y\u00FCzden tekrar yapman\u0131za gerek yok\!
+JPanelWebExport.jLabel3.text=1. Github Hesab\u0131n\u0131z\u0131n olmas\u0131 gerekmektedir.
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+CTL_PublishWeb=Web'de yay\u0131nla
+OpenIDE-Module-Long-Description=Bu eklenti a\u011F\u0131n\u0131z\u0131 bir gexf dosyas\u0131na d\u00F6n\u00FC\u015Ft\u00FCr\u00FCr. Bu dosya, herkesin eri\u015Fimine a\u00E7\u0131k hale geldi\u011Fi Github'daki bir "Gist"e g\u00F6nderilir. Ard\u0131ndan, a\u011F\u0131n\u0131z\u0131 Retina web uygulamas\u0131yla g\u00F6r\u00FCnt\u00FClemenize, ke\u015Ffetmenize ve payla\u015Fman\u0131za olanak tan\u0131yan bir ba\u011Flant\u0131 olu\u015Fturulur. Dikkatli olun, veriler gizliyse a\u011Flar\u0131n\u0131z\u0131 yay\u0131nlamay\u0131n\!
+JPanelWebExport.jLabel8.text=5. Orada bir kod g\u00F6r\u00FCnene kadar bekleyin ve kopyalay\u0131n -->;
+general.message.warning_confidentiality=\uD83D\uDD25Uyar\u0131\! Gizli veriler i\u00E7eriyorsa a\u011F\u0131n\u0131z\u0131 yay\u0131nlamay\u0131n\!\uD83D\uDD25
+general.message.error_message=Hata mesaj\u0131\:
+general.message.success_switch_to_publish=Vurmak. "Yay\u0131nla" se\u00E7ene\u011Fine gidebilirsiniz.
+general.message.url_published_gexf=A\u011F URL'niz\:
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. Github Hesab\u0131n\u0131za giri\u015F yap\u0131n
+general.message.info.url_will_appear_below=\uD83D\uDD17 A\u011F\u0131n\u0131z\u0131n ba\u011Flant\u0131s\u0131 (url) a\u015Fa\u011F\u0131da g\u00F6r\u00FCnecektir \uD83D\uDD17
+general.message.file_sent_from_gephi=Gephi'den g\u00F6nderilen dosya
+general.message.github.install_gephi_app=E\u011Fer yoksa, \u00FCcretsizdir ve burada olu\u015Fturmas\u0131 bir dakika s\u00FCrer\:
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=Github'a ba\u011Flan
+general.message.error.no_token=Yay\u0131nlanamad\u0131. Ayarlar\u0131 do\u011Fru yapt\u0131n\u0131z m\u0131?
+general.message.success_ready_to_publish=7. \u015Eimdi her \u015Fey yolunda olmal\u0131\! Bu pencerenin &quot;yay\u0131nla&quot; sekmesine gidin ve a\u011F\u0131n\u0131z\u0131 yay\u0131nlay\u0131n\!
+general.message.error.cant_retrieve_user_code=Kullan\u0131c\u0131 kodu al\u0131namad\u0131. Github'a d\u00FCzg\u00FCn bir \u015Fekilde ba\u011Fl\u0131 m\u0131s\u0131n\u0131z?
+general.message.error.no_open_project=Bir proje olu\u015Fturman\u0131z veya a\u00E7man\u0131z gerekir, aksi takdirde yay\u0131nlayacak bir \u015Fey yoktur.
+OpenIDE-Module-Name=A\u011F\u0131n\u0131z\u0131 web'de yay\u0131nlay\u0131n
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4. Gephi-lite uygulamas\u0131na ba\u011Flanmaya ba\u015Flay\u0131n, buraya t\u0131klay\u0131n\:
+JPanelWebExport.jLabel1.text=herhangi bir hata olu\u015Fursa, hata mesajlar\u0131 a\u015Fa\u011F\u0131da g\u00F6r\u00FCnecektir
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=Kurulum (sadece bir kez yapmak i\u00E7in)
+general.message.error.gexf_not_retrieved=A\u011F\u0131n\u0131zdan gexf dosyas\u0131 olu\u015Fturulam\u0131yor.
+general.message.success_reset=S\u0131f\u0131rlama ba\u015Far\u0131l\u0131 oldu.
+general.message.error.gist_creation=\u00C7evrimi\u00E7i gexf dosyas\u0131 olu\u015Fturulurken hata olu\u015Ftu. Hata kodu\:
+OpenIDE-Module-Short-Description=Bu eklenti, a\u011F\u0131n\u0131z\u0131 do\u011Frudan Gephi'den web \u00FCzerinde yay\u0131nlaman\u0131za, g\u00F6rselle\u015Ftirmenize ve payla\u015Fman\u0131za olanak tan\u0131r.
+general.message.error.empty_network=A\u011F\u0131n\u0131z bo\u015F. Yay\u0131nlamak i\u00E7in bir a\u011F olu\u015Fturun.
+general.message.plugin.name=Web'de yay\u0131nla
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=Giri\u015Fi s\u0131f\u0131rla / sil
+general.message.error.no_user_code=Hata. Kullan\u0131c\u0131 kodu web sitesinde do\u011Frulanmad\u0131.
+general.message.url_published_on_retina=A\u011F\u0131n\u0131z\u0131 do\u011Frudan Retina'da g\u00F6r\u00FCnt\u00FClemek i\u00E7in url\:
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=Yay\u0131nla
+general.verb.close_window=Kapat
+general.message.github.create_account=E\u011Fer yoksa, \u00FCcretsizdir ve burada olu\u015Fturmas\u0131 bir dakika s\u00FCrer\:
