+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_zh.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:17 CEST 2022
+JPanelWebExport.jButtonPublish.text=\u5C06\u60A8\u7684\u7F51\u7EDC\u53D1\u5E03\u5230\u7F51\u7EDC\uFF01
+general.message.github.past_code=6. \u5C06\u4EE3\u7801\u7C98\u8D34\u5230\u672C\u7AD9\uFF0C\u70B9\u51FB\u201C\u7EE7\u7EED\u201D\uFF0C\u518D\u70B9\u51FB\u201C\u6388\u6743\u201D\uFF1A
+general.message.error.unspecific_error_while_publishing=\u53D1\u5E16\u65F6\u51FA\u9519\uFF0C\u6CA1\u6709\u5177\u4F53\u7EC6\u8282\u3002
+general.message.warning_setup_already_done=\u770B\u8D77\u6765\u60A8\u5DF2\u7ECF\u5B8C\u6210\u4E86\u8BBE\u7F6E\uFF0C\u56E0\u6B64\u65E0\u9700\u518D\u6B21\u8BBE\u7F6E\uFF01
+JPanelWebExport.jLabel3.text=1.\u4F60\u9700\u8981\u6709\u4E00\u4E2AGithub\u8D26\u53F7\u3002
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+CTL_PublishWeb=\u53D1\u5E03\u5230\u7F51\u7EDC
+OpenIDE-Module-Long-Description=\u6B64\u63D2\u4EF6\u5C06\u60A8\u7684\u7F51\u7EDC\u8F6C\u6362\u4E3A gexf \u6587\u4EF6\u3002\u8BE5\u6587\u4EF6\u88AB\u53D1\u9001\u5230 Github \u7684\u201CGist\u201D\uFF0C\u5728\u90A3\u91CC\u5B83\u53EF\u4EE5\u516C\u5F00\u8BBF\u95EE\u3002\u7136\u540E\uFF0C\u4F1A\u521B\u5EFA\u4E00\u4E2A\u94FE\u63A5\uFF0C\u8BA9\u60A8\u53EF\u4EE5\u4F7F\u7528 Retina Web \u5E94\u7528\u7A0B\u5E8F\u67E5\u770B\u3001\u63A2\u7D22\u548C\u5171\u4EAB\u60A8\u7684\u7F51\u7EDC\u3002\u8BF7\u6CE8\u610F\uFF0C\u5982\u679C\u6570\u636E\u662F\u673A\u5BC6\u7684\uFF0C\u8BF7\u4E0D\u8981\u53D1\u5E03\u60A8\u7684\u7F51\u7EDC\uFF01
+JPanelWebExport.jLabel8.text=5. \u7B49\u5230\u90A3\u91CC\u51FA\u73B0\u4EE3\u7801\u5E76\u590D\u5236\u5B83-->;
+general.message.warning_confidentiality=\uD83D\uDD25\u8B66\u544A\uFF01\u5982\u679C\u60A8\u7684\u7F51\u7EDC\u5305\u542B\u673A\u5BC6\u6570\u636E\uFF0C\u8BF7\u4E0D\u8981\u53D1\u5E03\u5B83\uFF01\uD83D\uDD25
+general.message.error_message=\u9519\u8BEF\u4FE1\u606F\uFF1A
+general.message.success_switch_to_publish=\u6253\u3002\u4F60\u53EF\u4EE5\u53BB\u201C\u53D1\u5E03\u201D\u3002
+general.message.url_published_gexf=\u60A8\u7684\u7F51\u7EDC URL \u662F\uFF1A
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. \u767B\u5F55\u60A8\u7684 Github \u5E10\u6237
+general.message.info.url_will_appear_below=\uD83D\uDD17\u60A8\u7684\u7F51\u7EDC\u7684\u94FE\u63A5\uFF08url\uFF09\u5C06\u51FA\u73B0\u5728\u4E0B\u9762\uD83D\uDD17
+general.message.file_sent_from_gephi=\u4ECE Gephi \u53D1\u9001\u7684\u6587\u4EF6
+general.message.github.install_gephi_app=\u5982\u679C\u4F60\u6CA1\u6709\uFF0C\u5B83\u662F\u514D\u8D39\u7684\uFF0C\u9700\u8981\u4E00\u5206\u949F\u5728\u8FD9\u91CC\u521B\u5EFA\uFF1A
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=\u8FDE\u63A5\u5230 Github
+general.message.error.no_token=\u65E0\u6CD5\u53D1\u5E03\u3002\u60A8\u662F\u5426\u6B63\u786E\u8FDB\u884C\u4E86\u8BBE\u7F6E\uFF1F
+general.message.success_ready_to_publish=7. \u73B0\u5728\u4E00\u5207\u90FD\u5E94\u8BE5\u597D\u4E86\uFF01\u8F6C\u5230\u6B64\u7A97\u53E3\u7684\u201C\u53D1\u5E03\u201D\u9009\u9879\u5361\u5E76\u53D1\u5E03\u60A8\u7684\u7F51\u7EDC\uFF01
+general.message.error.cant_retrieve_user_code=\u65E0\u6CD5\u68C0\u7D22\u4ED6\u4EEC\u7684\u7528\u6237\u4EE3\u7801\u3002\u60A8\u662F\u5426\u6B63\u786E\u8FDE\u63A5\u5230 Github\uFF1F
+general.message.error.no_open_project=\u60A8\u5FC5\u987B\u521B\u5EFA\u6216\u6253\u5F00\u4E00\u4E2A\u9879\u76EE\uFF0C\u5426\u5219\u6CA1\u6709\u53EF\u53D1\u5E03\u7684\u5185\u5BB9\u3002
+OpenIDE-Module-Name=\u5C06\u60A8\u7684\u7F51\u7EDC\u53D1\u5E03\u5230\u7F51\u7EDC
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4.\u5F00\u59CB\u8FDE\u63A5Gephi-lite app\uFF0C\u70B9\u51FB\u8FD9\u91CC\uFF1A
+JPanelWebExport.jLabel1.text=\u5982\u679C\u53D1\u751F\u4EFB\u4F55\u9519\u8BEF\uFF0C\u9519\u8BEF\u6D88\u606F\u5C06\u51FA\u73B0\u5728\u4E0B\u65B9
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=\u8BBE\u7F6E\uFF08\u53EA\u505A\u4E00\u6B21\uFF09
+general.message.error.gexf_not_retrieved=\u65E0\u6CD5\u4ECE\u60A8\u7684\u7F51\u7EDC\u521B\u5EFA gexf \u6587\u4EF6\u3002
+general.message.success_reset=\u91CD\u7F6E\u6210\u529F\u3002
+general.message.error.gist_creation=\u521B\u5EFA\u5728\u7EBF gexf \u6587\u4EF6\u65F6\u51FA\u9519\u3002\u9519\u8BEF\u4EE3\u7801\uFF1A
+OpenIDE-Module-Short-Description=\u8BE5\u63D2\u4EF6\u5141\u8BB8\u60A8\u76F4\u63A5\u4ECE Gephi \u5728 Web \u4E0A\u53D1\u5E03\u3001\u53EF\u89C6\u5316\u548C\u5171\u4EAB\u60A8\u7684\u7F51\u7EDC\u3002
+general.message.error.empty_network=\u60A8\u7684\u7F51\u7EDC\u662F\u7A7A\u7684\u3002\u521B\u5EFA\u4E00\u4E2A\u7F51\u7EDC\u6765\u53D1\u5E03\u5B83\u3002
+general.message.plugin.name=\u53D1\u5E03\u5230\u7F51\u7EDC
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=\u91CD\u7F6E/\u5220\u9664\u767B\u5F55
+general.message.error.no_user_code=\u9519\u8BEF\u3002\u7528\u6237\u4EE3\u7801\u5C1A\u672A\u5728\u7F51\u7AD9\u4E0A\u9A8C\u8BC1\u3002
+general.message.url_published_on_retina=\u76F4\u63A5\u5728 Retina \u4E0A\u67E5\u770B\u60A8\u7684\u7F51\u7EDC\u7684\u7F51\u5740\u662F\uFF1A
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=\u53D1\u5E03
+general.verb.close_window=\u5173
+general.message.github.create_account=\u5982\u679C\u4F60\u6CA1\u6709\uFF0C\u5B83\u662F\u514D\u8D39\u7684\uFF0C\u9700\u8981\u4E00\u5206\u949F\u5728\u8FD9\u91CC\u521B\u5EFA\uFF1A
