+++ modules/WebPublishPlugin/src/main/resources/net/clementlevallois/web/publish/plugin/controller/Bundle_zh_TW.properties
@@ -1,0 +1,43 @@
+#Fri Sep 02 13:17:18 CEST 2022
+JPanelWebExport.jButtonPublish.text=\u5C07\u60A8\u7684\u7DB2\u7D61\u767C\u4F48\u5230\u7DB2\u7D61\uFF01
+general.message.github.past_code=6. \u5C07\u4EE3\u78BC\u7C98\u8CBC\u5230\u672C\u7AD9\uFF0C\u9EDE\u64CA\u201C\u7E7C\u7E8C\u201D\uFF0C\u518D\u9EDE\u64CA\u201C\u6388\u6B0A\u201D\uFF1A
+general.message.error.unspecific_error_while_publishing=\u767C\u5E16\u6642\u51FA\u932F\uFF0C\u6C92\u6709\u5177\u9AD4\u7D30\u7BC0\u3002
+general.message.warning_setup_already_done=\u770B\u8D77\u4F86\u60A8\u5DF2\u7D93\u5B8C\u6210\u4E86\u8A2D\u7F6E\uFF0C\u56E0\u6B64\u7121\u9700\u518D\u6B21\u8A2D\u7F6E\uFF01
+JPanelWebExport.jTextField1.text=https\://github.com/signup
+JPanelWebExport.jLabel3.text=1. \u4F60\u9700\u8981\u6709\u4E00\u500B Github \u8CEC\u865F\u3002
+CTL_PublishWeb=\u767C\u4F48\u5230\u7DB2\u7D61
+OpenIDE-Module-Long-Description=\u6B64\u63D2\u4EF6\u5C07\u60A8\u7684\u7DB2\u7D61\u8F49\u63DB\u70BA gexf \u6587\u4EF6\u3002\u8A72\u6587\u4EF6\u88AB\u767C\u9001\u5230 Github \u7684\u201CGist\u201D\uFF0C\u5728\u90A3\u88E1\u5B83\u53EF\u4EE5\u516C\u958B\u8A2A\u554F\u3002\u7136\u5F8C\uFF0C\u6703\u5275\u5EFA\u4E00\u500B\u93C8\u63A5\uFF0C\u8B93\u60A8\u53EF\u4EE5\u4F7F\u7528 Retina Web \u61C9\u7528\u7A0B\u5E8F\u67E5\u770B\u3001\u63A2\u7D22\u548C\u5171\u4EAB\u60A8\u7684\u7DB2\u7D61\u3002\u8ACB\u6CE8\u610F\uFF0C\u5982\u679C\u6578\u64DA\u662F\u6A5F\u5BC6\u7684\uFF0C\u8ACB\u4E0D\u8981\u767C\u5E03\u60A8\u7684\u7DB2\u7D61\uFF01
+JPanelWebExport.jLabel8.text=5. \u7B49\u5230\u90A3\u88E1\u51FA\u73FE\u4EE3\u78BC\u4E26\u8907\u5236\u5B83-->;
+general.message.warning_confidentiality=\uD83D\uDD25\u8B66\u544A\uFF01\u5982\u679C\u60A8\u7684\u7DB2\u7D61\u5305\u542B\u6A5F\u5BC6\u6578\u64DA\uFF0C\u8ACB\u4E0D\u8981\u767C\u5E03\u5B83\uFF01\uD83D\uDD25
+general.message.error_message=\u932F\u8AA4\u4FE1\u606F\uFF1A
+general.message.success_switch_to_publish=\u6253\u3002\u4F60\u53EF\u4EE5\u53BB\u201C\u767C\u5E03\u201D\u3002
+general.message.url_published_gexf=\u60A8\u7684\u7DB2\u7D61 URL \u662F\uFF1A
+JPanelWebExport.jTextFieldWebsiteLoginUrl.text=https\://github.com/login/device
+JPanelWebExport.jLabel5.text=2. \u767B\u9304\u60A8\u7684 Github \u5E33\u6236
+general.message.info.url_will_appear_below=\uD83D\uDD17\u60A8\u7684\u7DB2\u7D61\u7684\u93C8\u63A5\uFF08url\uFF09\u5C07\u51FA\u73FE\u5728\u4E0B\u9762\uD83D\uDD17
+general.message.file_sent_from_gephi=\u5F9E Gephi \u767C\u9001\u7684\u6587\u4EF6
+general.message.github.install_gephi_app=\u5982\u679C\u4F60\u6C92\u6709\uFF0C\u5B83\u662F\u514D\u8CBB\u7684\uFF0C\u9700\u8981\u4E00\u5206\u9418\u5728\u9019\u88E1\u5275\u5EFA\uFF1A
+JPanelWebExport.jTextField2.text=https\://github.com/apps/gephi-lite
+JPanelWebExport.jButtonGithubConnect.text=\u9023\u63A5\u5230 Github
+general.message.error.no_token=\u7121\u6CD5\u767C\u5E03\u3002\u60A8\u662F\u5426\u6B63\u78BA\u9032\u884C\u4E86\u8A2D\u7F6E\uFF1F
+general.message.success_ready_to_publish=7. \u73FE\u5728\u4E00\u5207\u90FD\u61C9\u8A72\u597D\u4E86\uFF01\u8F49\u5230\u6B64\u7A97\u53E3\u7684\u201C\u767C\u5E03\u201D\u9078\u9805\u5361\u4E26\u767C\u5E03\u60A8\u7684\u7DB2\u7D61\uFF01
+general.message.error.cant_retrieve_user_code=\u7121\u6CD5\u6AA2\u7D22\u4ED6\u5011\u7684\u7528\u6236\u4EE3\u78BC\u3002\u60A8\u662F\u5426\u6B63\u78BA\u9023\u63A5\u5230 Github\uFF1F
+general.message.error.no_open_project=\u60A8\u5FC5\u9808\u5275\u5EFA\u6216\u6253\u958B\u4E00\u500B\u9805\u76EE\uFF0C\u5426\u5247\u6C92\u6709\u53EF\u767C\u5E03\u7684\u5167\u5BB9\u3002
+OpenIDE-Module-Name=\u5C07\u60A8\u7684\u7DB2\u7D61\u767C\u4F48\u5230\u7DB2\u7D61
+JPanelWebExport.jTextFieldGithubErrorMsg.text=
+JPanelWebExport.jLabel7.text=4.\u958B\u59CB\u9023\u63A5Gephi-lite app\uFF0C\u9EDE\u64CA\u9019\u88E1\uFF1A
+JPanelWebExport.jLabel1.text=\u5982\u679C\u767C\u751F\u4EFB\u4F55\u932F\u8AA4\uFF0C\u932F\u8AA4\u6D88\u606F\u5C07\u51FA\u73FE\u5728\u4E0B\u65B9
+JPanelWebExport.tabGithub.TabConstraints.tabTitle=\u8A2D\u7F6E\uFF08\u53EA\u505A\u4E00\u6B21\uFF09
+general.message.error.gexf_not_retrieved=\u7121\u6CD5\u5F9E\u60A8\u7684\u7DB2\u7D61\u5275\u5EFA gexf \u6587\u4EF6\u3002
+general.message.success_reset=\u91CD\u7F6E\u6210\u529F\u3002
+general.message.error.gist_creation=\u5275\u5EFA\u5728\u7DDA gexf \u6587\u4EF6\u6642\u51FA\u932F\u3002\u932F\u8AA4\u4EE3\u78BC\uFF1A
+OpenIDE-Module-Short-Description=\u8A72\u63D2\u4EF6\u5141\u8A31\u60A8\u76F4\u63A5\u5F9E Gephi \u5728 Web \u4E0A\u767C\u5E03\u3001\u53EF\u8996\u5316\u548C\u5171\u4EAB\u60A8\u7684\u7DB2\u7D61\u3002
+general.message.error.empty_network=\u60A8\u7684\u7DB2\u7D61\u662F\u7A7A\u7684\u3002\u5275\u5EFA\u4E00\u500B\u7DB2\u7D61\u4F86\u767C\u5E03\u5B83\u3002
+general.message.plugin.name=\u767C\u4F48\u5230\u7DB2\u7D61
+JPanelWebExport.jTextFieldUserCode.text=XXX-XXX
+JPanelWebExport.jButtonResetLogin.text=\u91CD\u7F6E/\u522A\u9664\u767B\u9304
+general.message.error.no_user_code=\u932F\u8AA4\u3002\u7528\u6236\u4EE3\u78BC\u5C1A\u672A\u5728\u7DB2\u7AD9\u4E0A\u9A57\u8B49\u3002
+general.message.url_published_on_retina=\u76F4\u63A5\u5728 Retina \u4E0A\u67E5\u770B\u60A8\u7684\u7DB2\u7D61\u7684\u7DB2\u5740\u662F\uFF1A
+JPanelWebExport.tabPublish.TabConstraints.tabTitle=\u767C\u5E03
+general.verb.close_window=\u95DC
+general.message.github.create_account=\u5982\u679C\u4F60\u6C92\u6709\uFF0C\u5B83\u662F\u514D\u8CBB\u7684\uFF0C\u9700\u8981\u4E00\u5206\u9418\u5728\u9019\u88E1\u5275\u5EFA\uFF1A
